requirements.txt
flask
python-dotenv
flask_session
firebase_admin